Refactor: Update PersonRepository to use new helper classes
- Updated 'PersonRepository' to utilize 'PersonSqlStatements', 'PersonParameterBuilder', and 'PersonDataMapper'.
- Removes redundant code and improves adherence to SRP (Single Responsibility Principle).
- Enhances maintainability and readability by delegating responsibilities to dedicated classes.
